API change; tests updated and fixed.
Replaced MCBuilder.addWord with MCBuilder.addWordPair().  The latter takes
two arguments, the first and following word and registers them as a pair
of words in sequence.  This removes the previous word from the MCBuilder
and so allows the sequences to be constructed in any order.

Also updated builder_spec.rb to test the new API and fixed an unrelated
brainfart.